Cleaned up files. Moved player class to its own file and gave constants their own file
