+++ pyproject.toml
@@ -13,6 +13,7 @@ pytest = "^5.2"
 flake8 = "^4.0.1"
 black = "^22.1.0"
 pre-commit = "^2.17.0"
+isort = "^5.10.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
