+++ src/constants.py
@@ -1,0 +1,2 @@
+WIDTH = 1920
+HEIGHT = 1080
